requirements.txt
Flask==2.0.3
Jinja2==3.1.1
Flask-Babel>1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-Session==0.3.1
Flask-SocketIO==4.3.0
Flask-SQLAlchemy>2.5.0
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
gevent==1.4.0
greenlet==0.4.15
matplotlib==3.0.3
gunicorn==19.9.0
networkx==2.2
numpy==1.16.2
Pillow
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.7.3
python-decouple
python-engineio==3.13.0
python-socketio==4.6.0
pytz==2018.7
rope==0.12.0
scipy>=1.2.1
SQLAlchemy>=1.3.0
uuid==1.30
Werkzeug==2.0.0
WTForms==2.2.1
WTForms-SQLAlchemy==0.1
#cryptography

mysql-connector-python